All tests pass, all optimization routed through the optimizer object owned by context.  Next, to separate individual optimization states into separate strategies
