+++ src/compiler/compiler.rs
@@ -9,19 +9,14 @@ use clvm_rs::allocator::Allocator;
 
 use crate::classic::clvm::__type_compatibility__::{bi_one, bi_zero};
 use crate::classic::clvm_tools::stages::stage_0::TRunProgram;
-use crate::classic::clvm_tools::stages::stage_2::optimize::optimize_sexp;
 
-use crate::compiler::clvm::{convert_from_clvm_rs, convert_to_clvm_rs, sha256tree};
+use crate::compiler::clvm::sha256tree;
 use crate::compiler::codegen::{codegen, hoist_body_let_binding, process_helper_let_bindings};
-use crate::compiler::comptypes::{
-    CompileErr, CompileForm, CompilerOpts, DefunData, HelperForm, PrimaryCodegen,
-};
+use crate::compiler::comptypes::{CompileErr, CompileForm, CompilerOpts, PrimaryCodegen};
 use crate::compiler::dialect::{AcceptedDialect, KNOWN_DIALECTS};
-use crate::compiler::evaluate::{build_reflex_captures, Evaluator, EVAL_STACK_LIMIT};
 use crate::compiler::frontend::frontend;
-use crate::compiler::optimize::{deinline_opt, finish_optimization, NoOptimization};
+use crate::compiler::optimize::NoOptimization;
 use crate::compiler::prims;
-use crate::compiler::runtypes::RunFailure;
 use crate::compiler::sexp::{parse_sexp, SExp};
 use crate::compiler::srcloc::Srcloc;
 use crate::compiler::{BasicCompileContext, CompileContextWrapper};
@@ -106,66 +101,6 @@ pub fn create_prim_map() -> Rc<HashMap<Vec<u8>, Rc<SExp>>> {
     Rc::new(prim_map)
 }
 
-fn fe_opt(
-    allocator: &mut Allocator,
-    runner: Rc<dyn TRunProgram>,
-    opts: Rc<dyn CompilerOpts>,
-    compileform: CompileForm,
-) -> Result<CompileForm, CompileErr> {
-    let evaluator = Evaluator::new(opts.clone(), runner.clone(), compileform.helpers.clone());
-    let mut optimized_helpers: Vec<HelperForm> = Vec::new();
-    for h in compileform.helpers.iter() {
-        match h {
-            HelperForm::Defun(inline, defun) => {
-                let mut env = HashMap::new();
-                build_reflex_captures(&mut env, defun.args.clone());
-                let body_rc = evaluator.shrink_bodyform(
-                    allocator,
-                    defun.args.clone(),
-                    &env,
-                    defun.body.clone(),
-                    true,
-                    Some(EVAL_STACK_LIMIT),
-                )?;
-                let new_helper = HelperForm::Defun(
-                    *inline,
-                    DefunData {
-                        loc: defun.loc.clone(),
-                        nl: defun.nl.clone(),
-                        kw: defun.kw.clone(),
-                        name: defun.name.clone(),
-                        args: defun.args.clone(),
-                        orig_args: defun.orig_args.clone(),
-                        synthetic: defun.synthetic.clone(),
-                        body: body_rc.clone(),
-                        ty: defun.ty.clone(),
-                    },
-                );
-                optimized_helpers.push(new_helper);
-            }
-            obj => {
-                optimized_helpers.push(obj.clone());
-            }
-        }
-    }
-    let new_evaluator = Evaluator::new(opts.clone(), runner.clone(), optimized_helpers.clone());
-
-    let shrunk = new_evaluator.shrink_bodyform(
-        allocator,
-        Rc::new(SExp::Nil(compileform.args.loc())),
-        &HashMap::new(),
-        compileform.exp.clone(),
-        true,
-        Some(EVAL_STACK_LIMIT),
-    )?;
-
-    Ok(CompileForm {
-        helpers: optimized_helpers.clone(),
-        exp: shrunk,
-        ..compileform
-    })
-}
-
 fn do_desugar(program: &CompileForm) -> Result<CompileForm, CompileErr> {
     // Transform let bindings, merging nested let scopes with the top namespace
     let hoisted_bindings = hoist_body_let_binding(None, program.args.clone(), program.exp.clone())?;
@@ -184,42 +119,26 @@ fn do_desugar(program: &CompileForm) -> Result<CompileForm, CompileErr> {
     })
 }
 
-fn do_optimization_23(
-    context: &mut BasicCompileContext,
-    opts: Rc<dyn CompilerOpts>,
-    pre_forms: &[Rc<SExp>],
-) -> Result<SExp, CompileErr> {
-    let g = frontend(opts.clone(), pre_forms)?;
-    let desugared = do_desugar(&g)?;
-    let deinlined = deinline_opt(context, opts.clone(), desugared)?;
-    let generated = finish_optimization(&deinlined);
-    Ok(generated)
-}
-
 pub fn compile_pre_forms(
     context: &mut BasicCompileContext,
     opts: Rc<dyn CompilerOpts>,
     pre_forms: &[Rc<SExp>],
 ) -> Result<SExp, CompileErr> {
-    if opts.frontend_opt() && opts.dialect().stepping.map(|d| d > 22).unwrap_or(false) {
-        return do_optimization_23(context, opts, pre_forms);
-    }
-
     // Resolve includes, convert program source to lexemes
     let p0 = frontend(opts.clone(), pre_forms)?;
 
-    let p1 = if opts.frontend_opt() {
-        // Front end optimization
-        let runner = context.runner();
-        fe_opt(context.allocator(), runner, opts.clone(), p0)?
-    } else {
-        p0
-    };
+    let p1 = context.frontend_optimization(opts.clone(), p0)?;
 
     let p2 = do_desugar(&p1)?;
 
+    let p3 = context.post_desugar_optimization(opts.clone(), p2)?;
+
     // generate code from AST, optionally with optimization
-    codegen(context, opts, &p2)
+    let generated = codegen(context, opts.clone(), &p3)?;
+
+    let g2 = context.post_codegen_output_optimize(opts, generated)?;
+
+    Ok(g2)
 }
 
 pub fn compile_file(
@@ -237,28 +156,6 @@ pub fn compile_file(
         Box::new(NoOptimization::new()),
     );
     compile_pre_forms(&mut context_wrapper.context, opts, &pre_forms)
-}
-
-pub fn run_optimizer(
-    allocator: &mut Allocator,
-    runner: Rc<dyn TRunProgram>,
-    r: Rc<SExp>,
-) -> Result<Rc<SExp>, CompileErr> {
-    let to_clvm_rs = convert_to_clvm_rs(allocator, r.clone())
-        .map(|x| (r.loc(), x))
-        .map_err(|e| match e {
-            RunFailure::RunErr(l, e) => CompileErr(l, e),
-            RunFailure::RunExn(s, e) => CompileErr(s, format!("exception {e}\n")),
-        })?;
-
-    let optimized = optimize_sexp(allocator, to_clvm_rs.1, runner)
-        .map_err(|e| CompileErr(to_clvm_rs.0.clone(), e.1))
-        .map(|x| (to_clvm_rs.0, x))?;
-
-    convert_from_clvm_rs(allocator, optimized.0, optimized.1).map_err(|e| match e {
-        RunFailure::RunErr(l, e) => CompileErr(l, e),
-        RunFailure::RunExn(s, e) => CompileErr(s, format!("exception {e}\n")),
-    })
 }
 
 impl CompilerOpts for DefaultCompilerOpts {
